CSS files renamed, they don't work in templates. Add new tag function replaced from list.html to new-question.html
