+++ templates/display_a_question.html
@@ -2,7 +2,6 @@
 <html>
 
 <head>
-    <script src="https://use.fontawesome.com/51dda53a19.js"></script>
     <link rel="stylesheet" href="/static/display_a_question.css">
 </head>
